Swtich === to <=
If the application is structured such that it's possible to set up
a cyclical dependency, and a `take` stream emits a value that
synchronously results in another value being dispatched down
the `take` stream, the `this._n--` can get called twice, pushing it
down to -1 and thus never triggering `end`. This ensures the
stream will end after the value is emitted, even if another value
was emitted in the meantime.