Guarantee that subscription is available in handler
Previously, if an Observable emitted an event synchronously when the
observe() method was called (e.g. if it was a Property), the
Subscription object would not be available in the client scope.

For instance, this code would fail because the 'subscription' variable
is assigned after the handler completes:
subscription = obs.observe(v => subscription.unsubscribe())

This commit adds the subscription as a parameter to the 'onValue' and
'onError' handlers, making the object available to handlers without
breaking any existing usage.